# docs/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "cloacina-server API",
    "description": "REST API for the Cloacina workflow-orchestration server. WebSocket protocols (execution events, accumulator ingest, reactor commands, substrate delivery) are specified in the WebSocket Protocol reference of the documentation site.",
    "contact": {
      "name": "Dylan Storey",
      "email": "dylan.storey@gmail.com"
    },
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "0.7.0"
  },
  "paths": {
...
    "/v1/tenants/{tenant_id}/workflows": {
      "get": {
        "tags": [
          "workflows"
        ],
        "summary": "GET /tenants/:tenant_id/workflows — list registered workflows.",
        "operationId": "list_workflows",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "description": "Tenant identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Registered workflows",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantListResponse_WorkflowSummary"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Tenant access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "POST /tenants/:tenant_id/workflows — multipart upload of .cloacina source package.",
        "operationId": "upload_workflow",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "description": "Tenant identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "The first file field is taken as the .cloacina package",
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/PackageUploadForm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Package registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowUploadedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or empty package",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Tenant/role denied or signature verification failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/v1/tenants/{tenant_id}/workflows/{name}": {
      "get": {
        "tags": [
          "workflows"
        ],
        "summary": "GET /tenants/:tenant_id/workflows/:name — get workflow details.",
        "operationId": "get_workflow",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "description": "Tenant identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "Package name, or package UUID to inspect pending/building/failed rows",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workflow detail incl. build state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDetail"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Tenant access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Workflow not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/v1/tenants/{tenant_id}/workflows/{name}/{version}": {
      "delete": {
        "tags": [
          "workflows"
        ],
        "summary": "DELETE /tenants/:tenant_id/workflows/:name/:version — unregister workflow.",
        "operationId": "delete_workflow",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "description": "Tenant identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "Package name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version",
            "in": "path",
            "description": "Package version",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workflow unregistered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDeletedResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Tenant access or role denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Workflow not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "type": "object",
        "description": "Standardized error response body. Every non-2xx response from\n`cloacina-server` carries this shape; the request correlation ID is in\nthe `x-request-id` response header, not the body.",
        "required": [
          "error",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable error code (e.g. `invalid_pagination`,\n`workflow_not_found`)."
          },
          "error": {
            "type": "string",
            "description": "Human-readable error message."
          }
        }
      },
...
      "PackageUploadForm": {
        "type": "object",
        "description": "Multipart form for workflow package upload. Spec-only type: the handler\naccepts the first file field regardless of name; `file` is the\nconventional field name.",
        "required": [
          "file"
        ],
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "description": "The `.cloacina` package archive.",
            "contentMediaType": "application/octet-stream"
          }
        }
      },
...
      "TenantListResponse_WorkflowSummary": {
        "type": "object",
        "description": "List envelope variant that retains a top-level `tenant_id`, used by\ntenant-scoped list endpoints for backward compatibility with operator\ndashboards that key off it.",
        "required": [
          "tenant_id",
          "items",
          "total"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "One row in the workflow list (`GET /tenants/{tenant_id}/workflows`).",
              "required": [
                "id",
                "package_name",
                "version",
                "tasks",
                "created_at"
              ],
              "properties": {
                "created_at": {
                  "type": "string",
                  "description": "RFC 3339 timestamp."
                },
                "description": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "id": {
                  "type": "string",
                  "description": "Package UUID."
                },
                "package_name": {
                  "type": "string"
                },
                "tasks": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "description": "Task IDs included in this package."
                },
                "version": {
                  "type": "string"
                }
              }
            }
          },
          "tenant_id": {
            "type": "string"
          },
          "total": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "WorkflowDeletedResponse": {
        "type": "object",
        "description": "`DELETE /tenants/{tenant_id}/workflows/{name}/{version}` response.",
        "required": [
          "status",
          "package_name",
          "version"
        ],
        "properties": {
          "package_name": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "description": "Always `\"deleted\"`."
          },
          "version": {
            "type": "string"
          }
        }
      },
      "WorkflowDetail": {
        "type": "object",
        "description": "`GET /tenants/{tenant_id}/workflows/{name}` response — summary fields\nplus real build state (pending/building/failed/success).",
        "required": [
          "tenant_id",
          "id",
          "package_name",
          "version",
          "tasks",
          "created_at",
          "build_status"
        ],
        "properties": {
          "build_error": {
            "type": [
              "string",
              "null"
            ]
          },
          "build_status": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "description": "RFC 3339 timestamp."
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "description": "Package UUID."
          },
          "package_name": {
            "type": "string"
          },
          "tasks": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Task IDs included in this package."
          },
          "tenant_id": {
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        }
      },
...
      "WorkflowUploadedResponse": {
        "type": "object",
        "description": "`201 Created` body for a workflow package upload\n(`POST /tenants/{tenant_id}/workflows`, multipart).",
        "required": [
          "package_id",
          "tenant_id"
        ],
        "properties": {
          "package_id": {
            "type": "string",
            "description": "UUID of the registered package."
          },
          "tenant_id": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "http",
        "scheme": "bearer",
        "description": "Cloacina API key, sent as `Authorization: Bearer <key>`. Tenant scope is carried by the key itself; tenant-scoped routes additionally name the tenant in the path."
      }
    }
  },
...
}